+++ zellij-server/src/wasm_vm.rs
@@ -1,4 +1,4 @@
-use log::info;
+use log::{info, warn};
 use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::path::PathBuf;
@@ -28,8 +28,8 @@ use zellij_utils::{input::command::TerminalAction, serde, zellij_tile};
 
 #[derive(Clone, Debug)]
 pub(crate) enum PluginInstruction {
-    Load(Sender<u32>, PathBuf, usize), // tx_pid, path_of_plugin , tab_index
-    Update(Option<u32>, Event),        // Focused plugin / broadcast, event data
+    Load(Sender<u32>, PathBuf, usize, bool), // tx_pid, path_of_plugin , tab_index, allow_exec_host_cmd
+    Update(Option<u32>, Event),              // Focused plugin / broadcast, event data
     Render(Sender<String>, u32, usize, usize), // String buffer, plugin id, rows, cols
     Unload(u32),
     Exit,
@@ -54,6 +54,9 @@ pub(crate) struct PluginEnv {
     pub senders: ThreadSenders,
     pub wasi_env: WasiEnv,
     pub subscriptions: Arc<Mutex<HashSet<EventType>>>,
+    // FIXME: Once permission system is ready, this could be removed
+    pub _allow_exec_host_cmd: bool,
+    plugin_own_data_dir: PathBuf,
 }
 
 // Thread main --------------------------------------------------------------------------------------------------------
@@ -61,12 +64,15 @@ pub(crate) fn wasm_thread_main(bus: Bus<PluginInstruction>, store: Store, data_d
     info!("Wasm main thread starts");
     let mut plugin_id = 0;
     let mut plugin_map = HashMap::new();
+    let plugin_dir = data_dir.join("plugins/");
+    let plugin_global_data_dir = plugin_dir.join("data");
+    fs::create_dir_all(plugin_global_data_dir.as_path()).unwrap();
+
     loop {
         let (event, mut err_ctx) = bus.recv().expect("failed to receive event on channel");
         err_ctx.add_call(ContextType::Plugin((&event).into()));
         match event {
-            PluginInstruction::Load(pid_tx, path, tab_index) => {
-                let plugin_dir = data_dir.join("plugins/");
+            PluginInstruction::Load(pid_tx, path, tab_index, _allow_exec_host_cmd) => {
                 let wasm_bytes = fs::read(&path)
                     .or_else(|_| fs::read(&path.with_extension("wasm")))
                     .or_else(|_| fs::read(&plugin_dir.join(&path).with_extension("wasm")))
@@ -81,15 +87,15 @@ pub(crate) fn wasm_thread_main(bus: Bus<PluginInstruction>, store: Store, data_d
                     path.as_path().file_name().unwrap().to_str().unwrap(),
                     plugin_id,
                 );
+
+                let plugin_name = path.as_path().file_stem().unwrap();
+                let plugin_own_data_dir = plugin_global_data_dir.join(plugin_name);
+
                 let mut wasi_env = WasiState::new("Zellij")
                     .env("CLICOLOR_FORCE", "1")
-                    .preopen(|p| {
-                        p.directory(".") // FIXME: Change this to a more meaningful dir
-                            .alias(".")
-                            .read(true)
-                            .write(true)
-                            .create(true)
-                    })
+                    .map_dir("/host", ".")
+                    .unwrap()
+                    .map_dir("/data", plugin_own_data_dir.as_path())
                     .unwrap()
                     .stdin(Box::new(input))
                     .stdout(Box::new(output))
@@ -99,12 +105,18 @@ pub(crate) fn wasm_thread_main(bus: Bus<PluginInstruction>, store: Store, data_d
 
                 let wasi = wasi_env.import_object(&module).unwrap();
 
+                if _allow_exec_host_cmd {
+                    info!("Plugin({:?}) is able to run any host command, this may lead to some security issues!", path);
+                }
+
                 let plugin_env = PluginEnv {
                     plugin_id,
                     tab_index,
                     senders: bus.senders.clone(),
                     wasi_env,
                     subscriptions: Arc::new(Mutex::new(HashSet::new())),
+                    _allow_exec_host_cmd,
+                    plugin_own_data_dir,
                 };
 
                 let zellij = zellij_exports(&store, &plugin_env);
@@ -147,10 +159,16 @@ pub(crate) fn wasm_thread_main(bus: Bus<PluginInstruction>, store: Store, data_d
                     buf_tx.send(wasi_read_string(&plugin_env.wasi_env)).unwrap();
                 }
             }
-            PluginInstruction::Unload(pid) => drop(plugin_map.remove(&pid)),
+            PluginInstruction::Unload(pid) => {
+                info!("Bye from plugin {}", &pid);
+                // TODO: remove plugin's own data directory
+                drop(plugin_map.remove(&pid));
+            }
             PluginInstruction::Exit => break,
         }
     }
+    info!("wasm main thread exits");
+    fs::remove_dir_all(plugin_global_data_dir.as_path()).unwrap();
 }
 
 // Plugin API ---------------------------------------------------------------------------------------------------------
@@ -174,6 +192,7 @@ pub(crate) fn zellij_exports(store: &Store, plugin_env: &PluginEnv) -> ImportObj
         host_get_plugin_ids,
         host_open_file,
         host_set_timeout,
+        host_exec_cmd,
     }
 }
 
@@ -246,6 +265,24 @@ fn host_set_timeout(plugin_env: &PluginEnv, secs: f64) {
             ))
             .unwrap();
     });
+}
+
+fn host_exec_cmd(plugin_env: &PluginEnv) {
+    let mut cmdline: Vec<String> = wasi_read_object(&plugin_env.wasi_env);
+    let command = cmdline.remove(0);
+
+    // Bail out if we're forbidden to run command
+    if !plugin_env._allow_exec_host_cmd {
+        warn!("This plugin isn't allow to run command in host side, skip running this command: '{cmd} {args}'.",
+        	cmd = command, args = cmdline.join(" "));
+        return;
+    }
+
+    // Here, we don't wait the command to finish
+    process::Command::new(command)
+        .args(cmdline)
+        .spawn()
+        .unwrap();
 }
 
 // Helper Functions ---------------------------------------------------------------------------------------------------