+++ zellij-utils/src/input/layout.rs
@@ -53,9 +53,17 @@ pub enum SplitSize {
 #[serde(crate = "self::serde")]
 pub enum Run {
     #[serde(rename = "plugin")]
-    Plugin(Option<PathBuf>),
+    Plugin(Option<RunPlugin>),
     #[serde(rename = "command")]
     Command(RunCommand),
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone, PartialEq, Eq)]
+#[serde(crate = "self::serde")]
+pub struct RunPlugin {
+    pub path: PathBuf,
+    #[serde(default)]
+    pub _allow_exec_host_cmd: bool,
 }
 
 // The layout struct ultimately used to build the layouts.
@@ -95,9 +103,12 @@ impl LayoutFromYaml {
 
         let mut layout = String::new();
         layout_file.read_to_string(&mut layout)?;
-        let layout: LayoutFromYaml = serde_yaml::from_str(&layout)?;
+        let layout: Option<LayoutFromYaml> = serde_yaml::from_str(&layout)?;
 
-        Ok(layout)
+        match layout {
+            Some(layout) => Ok(layout),
+            None => Ok(LayoutFromYaml::default()),
+        }
     }
 
     // It wants to use Path here, but that doesn't compile.
@@ -216,6 +227,8 @@ pub struct TabLayout {
     pub parts: Vec<TabLayout>,
     pub split_size: Option<SplitSize>,
     pub run: Option<Run>,
+    #[serde(default)]
+    pub name: String,
 }
 
 impl Layout {
@@ -419,6 +432,7 @@ impl Default for TabLayout {
             parts: vec![],
             split_size: None,
             run: None,
+            name: String::new(),
         }
     }
 }